Refactor, reorganize, and cleanup.
