+++ app/src/util/helpers.js
@@ -1,0 +1,6 @@
+
+const { isArray } = Array;
+
+export const isNotStringOrStringArray = (thing) => (
+    typeof thing !== "string" && isArray(thing) && thing.some((p) => typeof p !== "string")
+);
